Move cross-platform functions, types, variables and constants to cross-platform code. GCW Zero port, part 2. It compiles and links, but may not work, and is intended for an early test of video only.
The following identifiers are now cross-platform: raise_interrupt, set_cpu_mode, prescale_table, GBA_SCREEN_WIDTH, GBA_SCREEN_HEIGHT, FIXED16_16, TIMER_TYPE, TIMER_IRQ_TYPE, TIMER_DS_CHANNEL_TYPE, TIMER_STATUS_TYPE.

+++ source/bios.c
@@ -20,8 +20,8 @@
 
 #include "common.h"
 
-unsigned int swi_hle_handle[0x2B][3] = {
-  /* { (unsigned int) void* HandlerAddress, unsigned int InputParameterCount, unsigned int ReturnsValue } */
+intptr_t swi_hle_handle[0x2B][3] = {
+  /* { (intptr_t) void* HandlerAddress, unsigned int InputParameterCount, unsigned int ReturnsValue } */
   { 0, 0, 0, },              // SWI 0:  SoftReset
   { 0, 0, 0, },              // SWI 1:  RegisterRAMReset
   { 0, 0, 0, },              // SWI 2:  Halt
